Asim/notion/requirements.txt
flask==3.0.2
python-dotenv==1.0.1
langchain==0.1.9
langchain-openai==0.0.7
langchain-google-genai==0.0.6
composio-langchain==latest_versio